optimized api

+++ backend/module/web_scrap/ColaManga/get.py
@@ -28,7 +28,7 @@ class RequestContextManager:
         global RequestQueueID
         while True:
             try:
-                RequestCache.objects.filter(room=RequestQueueRoom,client=self.id).delete()
+                if RequestQueueID:RequestCache.objects.filter(room=RequestQueueRoom,client=self.id).delete()
                 if RequestQueueID == self.id: RequestQueueID = None
                 break
             except Exception as e: print(f"Error Exiting Room: {RequestQueueRoom}.\n {e}\nRetrying...")
